Revert "Use conditional imports for flutter foundation libraries (#33663)" (#33825)
This reverts commit 90500a5dc4341dd41f3ae0d99fc92daa5ffd2a9c.
