+++ packages/flutter/lib/src/animation/animations.dart
@@ -3,6 +3,7 @@
 // found in the LICENSE file.
 
 import 'dart:math' as math;
+import 'dart:ui' show VoidCallback;
 
 import 'package:flutter/foundation.dart';
