+++ packages/flutter/lib/src/animation/listener_helpers.dart
@@ -2,6 +2,8 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
+import 'dart:ui' show VoidCallback;
+
 import 'package:flutter/foundation.dart';
 
 import 'animation.dart';
