--- packages/flutter/lib/src/foundation/_isolates_web.dart
@@ -1,14 +1,0 @@
-// Copyright 2019 The Chromium Authors. All rights reserved.
-// Use of this source code is governed by a BSD-style license that can be
-// found in the LICENSE file.
-
-import 'isolates.dart' as isolates;
-
-/// The dart:html implementation of [isolate.compute].
-Future<R> compute<Q, R>(isolates.ComputeCallback<Q, R> callback, Q message, { String debugLabel }) async {
-  // To avoid blocking the UI immediately for an expensive function call, we
-  // pump a single frame to allow the framework to complete the current set
-  // of work.
-  await null;
-  return callback(message);
-}
