+++ packages/flutter/lib/src/foundation/basic_types.dart
@@ -8,6 +8,7 @@ import 'dart:collection';
 // COMMON SIGNATURES
 
 export 'dart:ui' show VoidCallback;
+export 'bitfield.dart' if (dart.library.html) 'bitfield_unsupported.dart';
 
 /// Signature for callbacks that report that an underlying value has changed.
 ///
