+++ packages/flutter/lib/src/foundation/isolates.dart
@@ -3,9 +3,12 @@
 // found in the LICENSE file.
 
 import 'dart:async';
+import 'dart:developer' show Timeline, Flow;
+import 'dart:isolate';
 
-import '_isolates_io.dart'
-  if (dart.library.html) '_isolates_web.dart' as _isolates;
+import 'package:meta/meta.dart';
+
+import 'constants.dart';
 
 /// Signature for the callback passed to [compute].
 ///
@@ -16,9 +19,6 @@ import '_isolates_io.dart'
 ///
 /// {@macro flutter.foundation.compute.limitations}
 typedef ComputeCallback<Q, R> = FutureOr<R> Function(Q message);
-
-// The signature of [compute].
-typedef _ComputeImpl = Future<R> Function<Q, R>(ComputeCallback<Q, R> callback, Q message, { String debugLabel });
 
 /// Spawn an isolate, run `callback` on that isolate, passing it `message`, and
 /// (eventually) return the value returned by `callback`.
@@ -44,6 +44,82 @@ typedef _ComputeImpl = Future<R> Function<Q, R>(ComputeCallback<Q, R> callback, 
 ///
 /// The `debugLabel` argument can be specified to provide a name to add to the
 /// [Timeline]. This is useful when profiling an application.
-// Remove when https://github.com/dart-lang/sdk/issues/37149 is fixed.
-// ignore: prefer_const_declarations
-final _ComputeImpl compute = _isolates.compute;
+Future<R> compute<Q, R>(ComputeCallback<Q, R> callback, Q message, { String debugLabel }) async {
+  if (!kReleaseMode) {
+    debugLabel ??= callback.toString();
+  }
+  final Flow flow = Flow.begin();
+  Timeline.startSync('$debugLabel: start', flow: flow);
+  final ReceivePort resultPort = ReceivePort();
+  final ReceivePort errorPort = ReceivePort();
+  Timeline.finishSync();
+  final Isolate isolate = await Isolate.spawn<_IsolateConfiguration<Q, FutureOr<R>>>(
+    _spawn,
+    _IsolateConfiguration<Q, FutureOr<R>>(
+      callback,
+      message,
+      resultPort.sendPort,
+      debugLabel,
+      flow.id,
+    ),
+    errorsAreFatal: true,
+    onExit: resultPort.sendPort,
+    onError: errorPort.sendPort,
+  );
+  final Completer<R> result = Completer<R>();
+  errorPort.listen((dynamic errorData) {
+    assert(errorData is List<dynamic>);
+    assert(errorData.length == 2);
+    final Exception exception = Exception(errorData[0]);
+    final StackTrace stack = StackTrace.fromString(errorData[1]);
+    if (result.isCompleted) {
+      Zone.current.handleUncaughtError(exception, stack);
+    } else {
+      result.completeError(exception, stack);
+    }
+  });
+  resultPort.listen((dynamic resultData) {
+    assert(resultData == null || resultData is R);
+    if (!result.isCompleted)
+      result.complete(resultData);
+  });
+  await result.future;
+  Timeline.startSync('$debugLabel: end', flow: Flow.end(flow.id));
+  resultPort.close();
+  errorPort.close();
+  isolate.kill();
+  Timeline.finishSync();
+  return result.future;
+}
+
+@immutable
+class _IsolateConfiguration<Q, R> {
+  const _IsolateConfiguration(
+    this.callback,
+    this.message,
+    this.resultPort,
+    this.debugLabel,
+    this.flowId,
+  );
+  final ComputeCallback<Q, R> callback;
+  final Q message;
+  final SendPort resultPort;
+  final String debugLabel;
+  final int flowId;
+
+  R apply() => callback(message);
+}
+
+Future<void> _spawn<Q, R>(_IsolateConfiguration<Q, FutureOr<R>> configuration) async {
+  R result;
+  await Timeline.timeSync(
+    '${configuration.debugLabel}',
+    () async { result = await configuration.apply(); },
+    flow: Flow.step(configuration.flowId),
+  );
+  Timeline.timeSync(
+    '${configuration.debugLabel}: returning result',
+    () { configuration.resultPort.send(result); },
+    flow: Flow.step(configuration.flowId),
+  );
+}
