+++ packages/flutter/lib/src/foundation/platform.dart
@@ -2,8 +2,23 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import '_platform_io.dart'
-  if (dart.library.html) '_platform_web.dart' as _platform;
+import 'dart:io' show Platform;
+
+import 'assertions.dart';
+
+/// The platform that user interaction should adapt to target.
+///
+/// The [defaultTargetPlatform] getter returns the current platform.
+enum TargetPlatform {
+  /// Android: <https://www.android.com/>
+  android,
+
+  /// Fuchsia: <https://fuchsia.googlesource.com/>
+  fuchsia,
+
+  /// iOS: <http://www.apple.com/ios/>
+  iOS,
+}
 
 /// The [TargetPlatform] that matches the platform on which the framework is
 /// currently executing.
@@ -33,20 +48,36 @@ import '_platform_io.dart'
 // that would mean we'd be stuck with that platform forever emulating the other,
 // and we'd never be able to introduce dedicated behavior for that platform
 // (since doing so would be a big breaking change).
-TargetPlatform get defaultTargetPlatform => _platform.defaultTargetPlatform;
-
-/// The platform that user interaction should adapt to target.
-///
-/// The [defaultTargetPlatform] getter returns the current platform.
-enum TargetPlatform {
-  /// Android: <https://www.android.com/>
-  android,
-
-  /// Fuchsia: <https://fuchsia.googlesource.com/>
-  fuchsia,
-
-  /// iOS: <http://www.apple.com/ios/>
-  iOS,
+TargetPlatform get defaultTargetPlatform {
+  // TODO(jonahwilliams): consider where this constant should live.
+  const bool kIsWeb = identical(1, 1.0);
+  TargetPlatform result;
+  if (kIsWeb) {
+    result = TargetPlatform.android;
+  } else {
+    if (Platform.isIOS) {
+      result = TargetPlatform.iOS;
+    } else if (Platform.isAndroid) {
+      result = TargetPlatform.android;
+    } else if (Platform.isFuchsia) {
+      result = TargetPlatform.fuchsia;
+    }
+  }
+  assert(() {
+    if (!kIsWeb && Platform.environment.containsKey('FLUTTER_TEST'))
+      result = TargetPlatform.android;
+    return true;
+  }());
+  if (debugDefaultTargetPlatformOverride != null)
+    result = debugDefaultTargetPlatformOverride;
+  if (result == null) {
+    throw FlutterError(
+      'Unknown platform.\n'
+      '${Platform.operatingSystem} was not recognized as a target platform. '
+      'Consider updating the list of TargetPlatforms to include this platform.'
+    );
+  }
+  return result;
 }
 
 /// Override the [defaultTargetPlatform].
