+++ packages/flutter/lib/src/foundation/profile.dart
@@ -2,7 +2,8 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import 'basic_types.dart';
+import 'dart:ui' show VoidCallback;
+
 import 'constants.dart';
 
 /// DEPRECATED. `function` cannot be treeshaken out of release builds.
